requirements.txt
av==9.2.0
bokeh==2.4.3
colorama==0.4.6
colorcet==3.0.0
einops==0.4.1
holoviews==1.15.0
imageio==2.21.2
imgaug==0.4.0
joblib==1.1.0
matplotlib==3.5.3
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.3.5
Pillow==9.2.0
pyaml==21.10.1
pydicom==2.3.0
pytorchvideo==0.1.5
scikit-image==0.19.3
scikit-learn==1.0.2
scikit-optimize==0.9.0
scipy
seaborn==0.11.2
tensorboard==2.10.1
termcolor==2.0.1
timm==0.4.12
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.9.3
torchsummary==1.5.1
torchvision==0.14.0
tornado==6.2
tqdm==4.64.0
umap-learn==0.5.3
vidaug==1.5
vit-pytorch==0.35.8
numba==0.56.2
pretty-confusion-matrix==0.1.1
adan-pytorch==0.1.0